clean code (#10180)
* clean code

* clean code

* clean code

* clean code

+++ client/src/main/java/com/alibaba/nacos/client/naming/utils/Chooser.java
@@ -130,7 +130,7 @@ public class Chooser<K, T> {
          * Refresh.
          */
         public void refresh() {
-            Double originWeightSum = (double) 0;
+            double originWeightSum = 0;
             int size = 0;
             for (Pair<T> item : itemsWithWeight) {
